Switching to new pyvisa versions
Removed the visa intermediate module that was there to mediate for instruments that want to talk to a Prologix controller. In my opinion that should be done at the pyvisa level. I have updated two instrument wrappers for the new pyvisa style communication.

Committer: Gabriele de Boo <ggdeboo@gmail.com>